Added converters, fixed BaseEntity, fixed DTOs

+++ core/src/main/java/com/siemens/core/model/Broker.java
@@ -13,7 +13,7 @@ import java.util.Set;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class Broker extends BaseEntity<Long>{
+public class Broker extends BaseEntity<Integer>{
 
     @Column(name="name")
     private String name;
